requirements.txt
numpy==1.26.1
matplotlib==3.8.1
pyserial==3.5